+++ CHANGELOG.md
@@ -15,6 +15,6 @@ Critical items to know are:
 
 versions here coincide with releases on pypi.
 
-## [master](https://github.com/openbases/whedon-python/tree/master)
- - added base functions to get fields from paper markdown and tests  (0.0.1)
- - package registration (still under development)  (0.0.0)
+## [master](https://github.com/openbases/openbases-python/tree/master)
+ - added base functions to get fields from paper markdown and tests (0.0.0)
+ - package registration (still under development)